using filter like instead of in

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -15,7 +15,9 @@ if __name__ == '__main__':
     Base.metadata.create_all(engine)
     Session = sessionmaker(bind=engine)
     session = Session()
-    for row in session.query(State).filter(State.name.like('%a%')):
-        session.delete(row)
+    results = session.query(State).all()
+    for row in results:
+        if ('a' in row.name):
+            session.delete(row)
     session.commit()
     session.close()
